add textbox editing, scenes, doxygen comments

+++ UI/UISystem.cpp
@@ -1,11 +1,17 @@
 #include "UISystem.hpp"
 
-UISystem::UISystem(sf::Vector2f viewportSize) : viewportSize{viewportSize} {
+UISystem::UISystem(sf::Vector2i viewportSize) : viewportSize{viewportSize} {
   root = create_widget<Canvas>(nullptr);
   root->size = viewportSize;
   printf("Root: %p\n", root.get());
 }
 
+/**
+ * @brief Used to pass events into the UISystem to be distributed to the Widgets.
+ * 
+ * @param event The sf::Event we want to propagate
+ * @param mousePos The mouse position in game pixels when the event occured
+ */
 void UISystem::event_callback(const sf::Event &event,
                               const sf::Vector2i &mousePos) {
   if (root) {
@@ -13,6 +19,9 @@ void UISystem::event_callback(const sf::Event &event,
   }
 }
 
+/** 
+ * @brief Draw the whole tree onto the RenderTarget.
+ */
 void UISystem::draw(sf::RenderTarget &target, sf::RenderStates states) const {
   if (root) {
     target.draw(*root);
